Fix Clippy error regarding preview::preview module.
